+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Change: Update copyright date (2021)
 * Change: Tweak installation instruction
 * Unit tests:
+    * Change: Split shortcode-related tests out into their own file
     * Change: Restructure unit test file structure
         * Change: Move `phpunit/bin/` to `tests/bin/`
         * Change: Move `phpunit/bootstrap.php` into `tests/phpunit/`